pyproject.toml
[tool.poetry]
name = "fice-events-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "fice_events_bot" }]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = { version = "^3.0.0b7", allow-prereleases = true }
pydantic = { version = "^1.10.7", extras = ["dotenv"] }
dependencsqlalchemy = {version = "^2.0.9", extras = ["asyncio"]}
asyncpg = "^0.27.0"
alembic = "^1.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
